+++ NEWS.md
@@ -4,7 +4,9 @@ output: html_document
 ---
 
 
-* First version to be used by other packages, so bumping version to v0.1.0.
+* First version to be used by other packages, so bumping to v0.1.0.
+* Added new options for the `piece` argument of `get_piece()`: "pps", 
+  "prepositions", and "objects".
 * `RCLabels::prepositions` is now a vector instead of a list,
   thereby making downstream use of the object clearer.
 * Breaking changes